requirements.txt
Flask
Flask-Cors
requests
python-dotenv
Pillow
fpdf